requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
pinecone-client
pyttsx3
SpeechRecognition
typing-extensions
openai
